:wrench: added reindexing options to environment values

+++ src/index.js
@@ -18,20 +18,22 @@ module.exports = {
    * run jobs, or perform some special logic.
    */
   async bootstrap({ strapi }) {
-    algolia.init({
-      appId: process.env.ALGOLIA_APP_ID,
-      apiKey: process.env.ALGOLIA_API_KEY,
-    });
-    const posts = Array.from(
-      await strapi.db.query('api::post.post').findMany({
-        populate: {
-          tags: true,
-        },
-      })
-    ).map((p) => ({ ...p, objectID: p.id }));
-    console.log(`updating ${posts.length} post(s)...`);
-    // the Promise is left unwaited intentionally
-    await algolia.deleteObjects('posts');
-    algolia.saveObjects('posts', posts);
+    if (process.env.INDEXING_ON_BOOT === 'true') {
+      algolia.init({
+        appId: process.env.ALGOLIA_APP_ID,
+        apiKey: process.env.ALGOLIA_API_KEY,
+      });
+      const posts = Array.from(
+        await strapi.db.query('api::post.post').findMany({
+          populate: {
+            tags: true,
+          },
+        })
+      ).map((p) => ({ ...p, objectID: p.id }));
+      console.log(`updating ${posts.length} post(s)...`);
+      // the Promise is left unwaited intentionally
+      await algolia.deleteObjects('posts');
+      algolia.saveObjects('posts', posts);
+    }
   },
 };
